backend/vue_manager/serializers.py
from rest_framework import serializers
from people_manager.models import Person

class PersonSerializer(serializers.ModelSerializer):
    class Meta:
        model = Person
        fields = ('name', 'email')
