backend/requirements.txt
Django==2.0.4
django-cors-headers==2.2.0
djangorestframework==3.8.2
django-filter==1.1.0
psycopg2==2.7.4
